updated login page so skips quiz when quiz has been done before. Updated Explore table a bit (currently doesn't work correctly still fixing it)
TODO for me: fix the explore page. Currently the database doesn't work because for some reason the data inside the retrieveData() function (now in the viewDidLoad() function, is not allowing me to take the data outside of the snapshot closure. Probably an easy solution I am not thinking about right now. WIll finish tomorrow.

Also need to allow for adding of hobbies on table after that, and place that in the database.